Consolidate sql setup files
I was overthinking this by trying to keep the seed data that I need for
testing out of the production database. I wanted the production database
to have only 'good' data (i.e. no bogus users or reviews), as if this is
a real app. But for demonstration purposes it's better if it has some
seed data and an existing admin that can be used to test the authorized
routes.