Implement GET /api/v1/restaurants/search?query=*
Searches the name and cuisine columns using the 'ilike' operator. I
tried to use Postgres's "text seach" feature (sans indexing), but it
doesn't provide better results. I thought it would do some fuzzy
matching, but the only cool feature it seems to have aside from indexing
is that it normalizes lexemes. Unfortunately it doesn't consider
adjectival or demonymic forms for places to be from the same lexeme. So
'peru' doesn't match 'peruvian'... and that is no good for this
application.